# HCV-gDB/public/static/json/openapi_new.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "V-gDB",
    "description": "V-gDB is a virus database"
  },
  "servers": [
    {
      "url": "/rabv"
    }
  ],
  "tags":[
      {
          "name":"sequences",
          "display_name":"Sequences",
          "description":"API endpoints to lookup and retrieve sequences, alignments, meta-data."

      },
      {
          "name":"strains",
          "display_name":"Strains",
          "description":"API endpoints to lookup and strains and corrsponding segment data"

      },
      {
        "name":"phylogeny",
        "display_name":"Phylogeny",
        "description":"API endpoints to access phylogenetic trees"

      },      
      {
        "name":"analysis",
        "display_name":"Analysis",
        "description":"API endpoints to for various analysis modules (ie: adaptation mutations)"

      }
  ],
  "paths": {
    "/api/sequences/": {
      "path_display":"sequences/",
      "path_url":"api_sequences_metadata",
      "get": {
          "tags":["sequences"],
          "type":"GET",
        "summary": "Retrieve all NCBI sequences' meta-data",
        "description": "Retrieve all NCBI sequences' meta-data",
        "operationId": "get_all_ncbi_sequences__get",
        "parameters": [
          {
            "name": "gb_length_lower",
            "in": "query",
            "description":"Lower sequence length",
            "example":100,
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "gb_length_upper",
            "in": "query",
            "description":"Upper sequence length",
            "example":"400",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "gb_update_date_lower",
            "in": "query",
            "description":"Earliest year that the sequences have been updated in NCBI",
            "example":"2001",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "gb_update_date_upper",
            "in": "query",
            "description":"Latest year that the sequences have been updated in NCBI",
            "example":"2023",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "collection_year_lower",
            "in": "query",
            "description":"Earliest year that the sequences have been collected",
            "example":"1964",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "collection_year_upper",
            "in": "query",
            "description":"Latest year that the sequences have been collected",
            "example":"2023",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "creation_year_lower",
            "in": "query",
            "description":"Earliest year that the sequences have been created in NCBI",
            "example":"1964",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "creation_year_upper",
            "in": "query",
            "description":"Latest year that the sequences have been created in NCBI",
            "example":"2023",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "major_clade",
            "in": "query",
            "description":"An array of major clades",
            "example":"[Artic, Africa-2]",
            "required": false,
            "schema": {
              "type": "array"
            }
          },
          {
            "name": "minor_clade",
            "in": "query",
            "description":"Array of minor clades. If major clade is included, the minor clades must belong to the major clades.",
            "example":"[AL1a, AL2, AL3]",
            "required": false,
            "schema": {
              "type": "array"
            }
          },
          {
            "name": "host",
            "in": "query",
            "description":"Host species the sample was collected from",
            "example":"Canis familiaris",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "m49_region_id",
            "in": "query",
            "description":"Retrieve NCBI sequences that come from the corresponding region_id.",
            "example":"503",
            "external_link":"/get_m49_region_ids",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "name": "is_ldc",
            "in": "query",
            "description":"Is the sequence from a least developed country (LDC)",
            "example":"false",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "is_lldc",
            "in": "query",
            "description":"Is the sequence from a landlocked developing country (LLDC)",
            "example":"true",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "is_sids",
            "in": "query",
            "description":"Is the sequence from a small island developing state (SIDS)",
            "example":"true",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "outputs": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenbankEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sequence/{primary_accession}": {
      "path_display":"sequence/{primary_accession}",
      "path_url":"api_sequence_get_sequence_meta_data",
      "path_parameters":"sequence/",
      "get": {
          "tags":["sequences"],
          "type":"GET",
        "summary": "Lookup NCBI primary_accession and retrieve corresponding meta-data.",
        "description": "Retrieve NCBI sequence with specific primary_accession",
        "operationId": "get_all_ncbi_sequences__get",
        "parameters": [
          {
            "name": "primary_accession",
            "description":"NCBI primary accession",
            "in": "query",
            "required": true,
            "example":"MT862689",
            "schema": {
              "type": "string"
            }
          }
        ],
        "outputs": [
                    {
                      "name":"meta_data",
                      "description": "a dictionary of meta-data available"
                    },
                    {
                      "name":"sequence",
                      "description": "the unaligned sequence"
                    },
                    {
                      "name":"regions",
                      "description": "a dictionary of regional data about the sample"
                    },
                    {
                      "name":"insertions",
                      "description": "a list of the nucleotide positions and corresponding insertions for the aligned sequence"
                    },
                    {
                      "name":"alignment",
                      "description": "a dictionary containing the aligned sequence, the reference accession and reference aligned sequence, and the features"
                    }
                  ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenbankEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sequence/reference/{primary_accession}": {
      "path_display":"/api/alignment/get_reference_sequences_meta_data/",
      "path_url":"api_alignment_get_reference_sequences",
      "get": {
          "tags":["sequences"],
          "type":"GET",
        "summary": "Lookup NCBI primary_accession for reference sequence and retrieve reference data",
        "description": "Retrieve reference sequence data based on reference NCBI primary_accession",
        "operationId": "get_all_ncbi_sequences__get",
        "parameters": [],
        "outputs": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenbankEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sequences/global/": {
      "path_display":"sequences/global/",
      "path_url":"api_get_statistics",
      "get": {
          "tags":["sequences"],
          "type":"GET",
        "summary": "Retrieve counts of sequences per country",
        "description": "Retrieve counts of sequences per country",
        "operationId": "get_sequences_global__get",
        "parameters": [],
        "outputs": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenbankEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/phylogeny/tree/": {
      "path_display":"phylogeny/tree/",
      "path_url":"api_get_phylogeny_tree",
      "get": {
          "tags":["phylogeny"],
          "type":"GET",
        "summary": "Retrieve phylogenetic tree",
        "description": "Retrieve phylogenetic tree",
        "operationId": "get_phylogeny_tree__get",
        "parameters": [],
        "outputs": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenbankEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/clade_assignment/": {
      "path_display":"analysis/clade_assignment/",
      "path_url":"api_analysis_clade_assignment",
      "get": {
          "tags":["analysis"],
          "type":"GET",
        "summary": "Run phylogenetic clade assignment analysis module",
        "description": "Run phylogenetic clade assignment analysis module",
        "operationId": "analysis_clade_assignment__get",
        "parameters": [{
            "name": "fasta",
            "in": "query",
            "description":"fasta sequence in the format of >accession with the amino acid sequence seperated by newline",
            "example":"",
            "required": true,
            "schema": {
              "type": "text"
            }
          }],
        "outputs": [{
                      "name":"accession",
                      "description": "the accession of the submitted sequence"
                    },
                  {
                      "name":"alignment",
                      "description": "the submitted sequence aligned to the closest reference"
                    },
                  {
                      "name":"blast_identity",
                      "description": "the percentage of BLAST result"
                    },
                  {
                      "name":"blast_ref",
                      "description": "the reference chosen by BLAST search"
                    },
                  {
                      "name":"epa_ng",
                      "description": "clade assignment"
                    },
                  {
                      "name":"tree",
                      "description": "newick tree with submitted sequence placed on reference tree"
                    }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenbankEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }      
  }
}